资源重构步骤1: 更新CMake配置以使用64位EnTT资源/哈希ID，降低冲突风险。修改多个调试面板以支持64位ID显示，确保一致性和准确性。调整保存数据结构以使用64位ID，增强数据兼容性。

+++ src/game/save/save_data.h
@@ -43,7 +43,7 @@ struct GameTimeSaveData {
 };
 
 struct ItemStackSaveData {
-    std::uint32_t item_id{0};
+    std::uint64_t item_id{0};
     int count{0};
 };
 
@@ -85,7 +85,7 @@ struct ResourceNodeSaveData {
     std::string node_type{"unknown"};
     int hit_count{0};
     int hits_to_break{0};
-    std::uint32_t drop_item_id{0};
+    std::uint64_t drop_item_id{0};
 };
 
 struct MapSaveData {
