+++ src/game/system/item_use_system.cpp
@@ -169,7 +169,7 @@ void ItemUseSystem::onUseItem(const game::defs::UseItemCommand& evt) {
             const int add_total = effect.count * use_times;
             if (effect.item_id == entt::null || add_total <= 0) continue;
             const auto* out_item = catalog_.findItem(effect.item_id);
-            const std::string name = out_item ? out_item->display_name_ : std::to_string(static_cast<std::uint32_t>(effect.item_id));
+            const std::string name = out_item ? out_item->display_name_ : std::to_string(static_cast<unsigned long long>(effect.item_id));
             if (!first) {
                 oss << "\n";
             }
